check existence of file/directory in clean_repo.sh

+++ clean_repo.sh
@@ -2,7 +2,16 @@
 script_path=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
 cd "$script_path"
 
-rm -r Java
-rm log_files/log_file.txt
+if [ -d  Java ]
+then
+   rm -r Java
+   echo "folder deleted: language_studier/Java"
+fi
+
+if [ -f  log_files/log_file.txt ]
+then
+   rm log_files/log_file.txt
+   echo "file deleted: language_studier/log_files/log_file.txt"
+fi
 
 #data/settings_data/* is not cleaned
